requirements.txt
beautifulsoup4
requests
dateparser
apache-airflow-providers-mongo
pymongo==3.6.0
pandas